+++ gehomesdk/erd/converters/common/__init__.py
@@ -1,4 +1,5 @@
 from .erd_demand_response_state_converter import ErdDemandResponseStateConverter
 from .erd_dsm_power_usage_converter import ErdDsmPowerUsageConverter
 from .erd_gas_type_converter import ErdGasTypeConverter
+from .erd_cumulative_gas_converter import ErdCumulativeGasCubicFeetConverter
 from .erd_resource_measurement_type_converter import ErdResourceMeasurementTypeConverter